feat: update friends management with new AddFriendTab and Friends components, remove deprecated files Fixes #75
