+++ client/src/app/router/Routes.tsx
@@ -13,7 +13,7 @@ import AuthCallback from "../../features/account/AuthCallback";
 import ChatRoomDetails from "../../features/chatRooms/details/ChatRoomDetails";
 import JoinChatRoomPage from "../../features/chatRooms/join/JoinChatRoomPage";
 import ProfilePage from "../../features/profile/ProfilePage";
-import FriendsList from "../../features/friends/FriendsList";
+import Friends from "../../features/friends/Friends";
 import AuthLayout from "../layout/AuthLayout";
 import EmptyPage from "../layout/EmptyPage";
 import DirectChatDetails from "../../features/directChats/DirectChatDetails";
@@ -30,7 +30,7 @@ export const router = createBrowserRouter([
           { path: "chat-rooms/:id", element: <ChatRoomDetails /> },
           { path: "chat-rooms/:id/:token/join", element: <JoinChatRoomPage /> },
           { path: "profiles/:id", element: <ProfilePage /> },
-          { path: "friends", element: <FriendsList /> },
+          { path: "friends", element: <Friends /> },
           { path: "direct-chats/:id", element: <DirectChatDetails /> },
           { path: "not-found", element: <NotFound /> },
           { path: "server-error", element: <ServerError /> },
