+++ client/src/features/chatRooms/ChatRoomSidebarContent.tsx
@@ -83,13 +83,13 @@ export default function ChatRoomSidebarContent() {
           sx={{
             cursor: "pointer",
             fontWeight: "bold",
-            bgcolor: "#2b2d31",
+            bgcolor: "#1e1f24",
             color: "white",
             px: 2,
             py: 1,
+            width: "100%",
             borderRadius: 2,
-            "&:hover": { bgcolor: "#36373c" },
-            width: "fit-content",
+            "&:hover": { bgcolor: "#2e2f33ff" },
             display: "flex",
             alignItems: "center",
             gap: 1,
